backend/requirements.txt
uvicorn[standard]
fastapi==0.88.0
sqlmodel==0.0.8
python-dotenv
psycopg2-binary==2.9.5
python-jose[cryptography]
passlib[bcrypt]
python-multipart
redis==4.4.0